api: update device type once available
The SurviveSimpleObject_type was never set to HMD, because to figure out
if a device is an HMD, the config parsing needed to be triggered from
the usb polling.

Now the type will be kept at unknown until it can be updated to an
actual type. The application will have to wait for the type to be
available before knowing which device is which.